tests/wpt/web-platform-tests/css/css-align/parsing/justify-self-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Box Alignment Level 3: getComputedStyle().justifySelf</title>
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-justify-self">
<meta name="assert" content="justify-self computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("justify-self", "auto");
test_computed_value("justify-self", "normal");
test_computed_value("justify-self", "stretch");

test_computed_value("justify-self", "baseline");
test_computed_value("justify-self", "last baseline");

test_computed_value("justify-self", "center");
test_computed_value("justify-self", "start");
test_computed_value("justify-self", "end");
test_computed_value("justify-self", "self-start");
test_computed_value("justify-self", "self-end");
test_computed_value("justify-self", "flex-start");
test_computed_value("justify-self", "flex-end");
test_computed_value("justify-self", "unsafe center");
test_computed_value("justify-self", "safe self-end");
test_computed_value("justify-self", "left");
test_computed_value("justify-self", "unsafe right");
</script>
</body>
</html>
